Opdracht e8

+++ E6.py
@@ -4,11 +4,11 @@ import pcinput
 # Inputs
 oorspronkelijkSaldo = pcinput.getFloat('Geef oorspronkelijk saldo: ')
 rente = pcinput.getFloat('Geef rente: ')
-aantalJaar = pcinput.getFloat('Geef het aantal jaar: ')
+aantalJaar = pcinput.getInteger('Geef het aantal jaar: ')
 
 # Berekeningen
 groeifactor = 1 + rente/100
 nieuwSaldo = round(oorspronkelijkSaldo * groeifactor**aantalJaar,2)
 
 # Eindresultaat printen
-print('Na {} jaar is je banksaldo {}').format(int(aantalJaar), nieuwSaldo)
+print('Na {} jaar is je banksaldo {}').format(aantalJaar, nieuwSaldo)
